Add dates to simple return vector
Add dates to simple return vector

The vector sbux_ret now contains the simple returns of Starbucks (well done!). It would be convenient to have the dates as names for the elements of that vector. Remember that the trading dates were in the first column of the sbux_df data frame. To set the names of a vector, you can use names(vector) <- some_names.

Remember that we are dealing with closing prices. The first return in sbux_df is thus realized on the second day, or sbux_prices_df[2,1].
Instructions

    Assign the correct dates as names to all elements of the return vector sbux_ret.
    Use the head() function to display the first elements of sbux_ret.