Compare simple and continuously compounded returns
You would like to compare the simple and the continuously compounded returns. In the next exercise, you will do that by generating two graphs. In this exercise, you will just have a quick look at the data. It would be nice to have the simple and continuously compounded return next to each other in a matrix, with n rows and two columns. You can use the cbind() function to paste the two vectors that contain both types of returns next to each other in a matrix.
Instructions

    Use the cbind() and head() functions to display the simple and continuously compounded returns next to each other for the first time periods. What do you notice?