Compute probabilities
In this lab you will become more familiar with random variables and probability distributions.

Suppose X is a normally distributed random variable with a mean 0.05 and a variance (0.10)2. Start by computing some probabilities.

Probabilities for normally distributed random variables can be computed with the function pnorm(). For instance, if Z is a standard normal random variable, Pr(Z≤0.5) can be computed with pnorm(0.5). You can supply the mean and the standard deviation of the normal distribution via the arguments mean and sd, respectively.
Instructions

    Assign to mu_x the mean of X.
    Assign to sigma_x the standard deviation of X.
    Compute the following probabilities and print the result to the console:
        Pr(X > 0.10)
        Pr(X < -0.10)
        Pr(-0.05 < X < 0.15)