Compute quantiles
Continue with computing quantiles of the normally distributed random variable X with a mean 0.05 and a variance (0.10)2.

Quantiles of normally distributed random variables can be computed with the function qnorm(). For instance, if Z is a standard normal random variable, the 10% quantile can be computed with qnorm(0.1). You can supply the mean and the standard deviation of the normal distribution via the arguments mean and sd, respectively.
Instructions

    Compute the 1%, 5%, 95% and 99% quantile with a single command. Do not assign it to a variable, just print it to the console.