Determine the value-at-risk of simple monthly returns
Consider again the Microsoft stock. Assume that the simple monthly return R is normally distributed with a mean 0.04 and a variance (0.09)2. The initial wealth to be invested over the month is $100,000.

Determine the 1% and the 5% value-at-risk (VaR) over the month on the investment. That is, determine the loss in investment value that may occur over the next month with a 1% probability and with a 5% probability.
Instructions

    Assign to mu_R the mean of R.
    Assign to sigma_R the standard deviation of R.
    Assign to W0 the initial wealth.
    Compute the 1% value-at-risk and print the result to the console.
    Compute the 5% value-at-risk and print the result to the console.