Determine the value-at-risk of continuously compounded monthly returns
Instead of the simple monthly return, now look at the continuously compounded monthly return r of the Microsoft stock. Assume that r is normally distributed with a mean 0.04 and a variance (0.09)2. The initial wealth to be invested over the month is $100,000.

Determine the 1% and the 5% value-at-risk (VaR) over the month on the investment. That is, determine the loss in investment value that may occur over the next month with a 1% probability and with a 5% probability.

Use the fact that the continuously compounded return quantile can be transformed to a simple return quantile with the transformation R=er−1. The exponential er can easily be computed with exp(r).
Instructions

    Assign to mu_r the mean of r.
    Assign to sigma_r the standard deviation of r.
    Assign to W0 the initial wealth.
    Compute the 1% value-at-risk and print the result to the console.
    Compute the 5% value-at-risk and print the result to the console.