Compute a joint probability
The red line segments in the scatterplot on the right indicate the area where X≤0 and Y≤0. Your task is to compute the joint probability Pr(X≤0,Y≤0).

Probabilities for a bivariate (or more generally, a multivariate) normal distribution can be computed with the function pmvnorm from package mvtnorm (see its documentation).
You should supply the lower limits as the lower argument and the upper limits as the upper argument. Note that you can use -Inf for −∞ and Inf for ∞. In addition, you can set the mean vector and the covariance matrix via the arguments mean and sigma, respectively.

Note that you can use args(pmvnorm) to display the arguments of the pmvnorm() function. Try this out in the console.
Instructions

    Compute the joint probability Pr(X≤0,Y≤0).