Plotting the theoretical and the sample ACF
Still consider the MA(1) model
Yt=0.05+ϵt+θϵt−1,
with |θ|<0 and ϵt iid N(0,(0.1)2).

While we still assume θ=0.5, the theoretical value for the first order autocorrelation is ρ1=θ1+θ2=0.51+0.52=0.4. The theoretical autocorrelation function gives you for each lag the autocorrelation implied by the model. In R, you can use the ARMAacf function to calculate the theoretical autocorrelations for an ARMA model. The ma argument should contain the moving average coefficients, and lag.max specifies the maximum number of lags for which the autocorrelations should be calculated.

Based on the simulated data, you can also easily calculate the sample autocorrelation function in R with the acf() function. It automatically generates a plot of the sample autocorrelation function. Check the documentation of acf for more info.
Instructions

    Use the ARMAacf function to calculate the theoretical autocorrelations up to lag 10. Assign the result to acf_ma1_model.
    Construct the plot of the sample autocorrelations up to lag 10 by using the acf() function. Set the title of the plot to "Sample ACF".