Getting the financial data
In this lab, you will learn more about the analysis of stock returns by using the R packages PerformanceAnalytics, zoo and tseries (click them for more information).

In this first exercise, you will use the get.hist.quote function from the tseries package. The function allows you to download historical financial data over the web. The code on the right downloads the adjusted closing prices from the start of 1998 up to the end of 2009 from Yahoo! for the stocks with tickers VBLTX and FMAGX. The code further specifies that the data should be of the zoo class which is indexed by time. Since we deal with monthly data, it is a good idea to change the class of the time series index with the as.yearmon() function.
Instructions

    Assign to SBUX_prices the prices of the Starbucks stock in the same way as VBLTX and FMAGX.
    Change the class of the index of SBUX_prices with the as.yearmon() function.
    Print the start and the end of the time series to the console by applying the start() and end() functions on one of the price time series.