Calculating the returns
In the previous exercise, we obtained the price series for multiple stocks. It is convenient to store these time series in a single variable before we start the analysis. You will use the merge() function to do that, since it merges series by column while taking into account the time index.

Remember that the continuously compounded returns are defined as the difference between the log prices. Once all price series are merged by column, you can easily calculate the continuously compounded returns. Use the log() function to calculate the log prices and apply the diff() function to the object that contains the log prices to get the continuously compounded returns.
Instructions

    Assign to all_prices the merged price series VBLTX_prices, FMAGX_prices, SBUX_prices such that each column contains the prices of one stock.
    Calculate the continuously compounded returns with the help of the diff() and log() functions.
    Have a look at the data in the console.