Create graphical summary for a return series
As a first step, the code on the right extracts the "core data" from the all_returns variable and stores it in the matrix return_matrix. "Extracting the core data" in this case means stripping off the index/time attributes and returning only the observations. Secondly, the graphical window is subdivided into two rows and two columns with the par(mfrow=c(2,2,)) command. Finally, the code generates a histogram, boxplot, density and qqplot of the return data. This allows you to visually assess the symmetry of the return distribution, whether or not there are fat tails, outliers, etc.
Instructions

    Click 'Submit Answer' to inspect the monthly returns of VBLTX.