Estimate the standard error of the correlation parameter
Still consider the constant expected return model (CER) introduced in the previous exercise. Correlations indicate the strength of the dependency between two variables. You are now interested in estimates of the correlations ρij between the returns of the different assets i and j. Furthermore, you would like to investigate the precision of these estimates by calculating the standard error of ρ^ij, which will be used for inference in later exercises.

Recall that the estimated SE values are computed using the analytic formula:
SE^(ρij^)=(1−ρij^2)/T√.
Instructions

    Assign to cor_matrix the correlation matrix of the returns.
    Assign to rhohat_vals the estimates of the correlations between "VBLTX,FMAGX","VBLTX,SBUX","FMAGX,SBUX".
    Assign to se_rhohat the estimates of SE(ρ2ij^) and print the result to the console.