Normality of the asset returns
Remember that the Constant Expected Return model assumes that returns are normally distributed. Is that a reasonable assumption? The Jarque Bera test provides a way to answer that question. You can easily perform the Jarque Bera test for normality in R with the jarque.bera.test function.

Let us say that you want to investigate whether it is reasonable that the returns of VBLTX are normally distributed. More formally, you would like to test the null hypothesis:

H0:rit∼ normal vs. H1:rit∼ not normal 

using a 5% significance level, with i=VBLTX.
Instructions

    Use the jarque.bera.test to test the normality of the "VBLTX" returns and print the output to the console. What do you conclude?