Tangency portfolio and T-bills
In the previous exercise, you calculated the composition of the tangency portfolio. As a second step, you create new portfolios that are a combination of T-bills and this tangency portfolio. First, you will need to compute the expected portfolio return and volatility for these portfolios. Second, your job is to create a plot of these portfolios with the volatility on the x-axis and the expected return on the y-axis.
Instructions

    Calculate the expected portfolio returns mu_portfolio_tangency_bill and portfolio volatilities sigma_portfolio_tangency_bill for the set of weights and indicate the proportion that was invested in the tangency portfolio (tangency_weights).
    Add these new portfolios to the plot that you constructed in the previous exercise with the help of the points() function. For the color code argument use col="blue", for the type use type="b", and as a plotting character use pch=16.