An Efficient Portfolio with 30% Tangency
Consider a portfolio that has 30% invested in the tangency portfolio and 70% in T-bills. Before making the final call, you want to investigate this portfolio's return and volatility in more detail. You decide to use the techniques that you learned in the previous exercises to map the portfolio in the spectrum of available portfolios.
Instructions

    Set the tangency_weight to 0.3 and then determine the t_bill_weight.
    Assign to mu_portfolio_efficient the expected portfolio return and to sd_portfolio_efficient the portfolio standard deviation.
    Plot this portfolio using points. Use “orange”“ as the color for the line, the type is "b”, the plotting character is 16, and apply a character expansion cex of 2. The corresponding text() function is already provided.