The global minimum variance portfolio - Part Two
Now you want to construct the global minimum variance portfolio under the condition that short sales are not allowed. The Markowitz portfolio optimization problem for the minimum variance portfolio with no short sales restrictions can be described as follows:

minx σ2p,x=x′∑x, subject to
x′1=1
xi≥0,

with x the vector of portfolio weights, σ2p,x and μp,x the portfolio variance and expected return respectively, μ the vector of expected returns and Σ the covariance matrix of the returns.

As seen in the lectures, the portfolio optimization problem with inequality constraints can be set up as a quadratic programming problem. Quadratic programming problems are of the form:
minx 12x′Dx−d′x, subject to
A′neqx≥bneq, for m inequality constraints, and 
A′eqx=bneq, for l equality constraints,
where D is a n×n matrix, x and d are n×1 vectors, A′neq is an m×n matrix, bneq is an m×1 vector, A′eq is an l×n matrix, and beq is an l×1 vector.

Quadratic programming problems can be solved with the R package quadprog and the solve.QP() function. In the next exercise you will discover a more easy way to do this with the help of globalMin.portfolio().
Instructions

    The restriction matrices are already set up. Make sure to study their output.
    Use these restriction matrices and the global minimum variance portfolio equality constraint to minimize the portfolio variance via solve.QP(). Assign the result to quad_prog.