The tangency portfolio
Let us see how the weights of the underlying assets of the tangency portfolio depend on whether or not short positions are allowed.

Assume the risk-free rate is 0.005 (rf = 0.5%) per month. The tangency portfolio can be found via:

maxt slope=μp−rfσp, subject to
μp=t′μ
σp=(t′∑t)½
t′1=1,

with μp and σp the portfolio return and standard deviation respectively, t the vector of portfolio weights, μ the vector of expected returns and Σ the covariance matrix of the returns.

If you add the condition that no short positions are allowed, the additional constraint that is to be added is xi≥ for i=1,…,4.

Luckily, the underlying arithmetics are encapsulated in the R function tangency.portfolio(). Again, smart use of these functions can make your work as a financial analyst considerably lighter.
Instructions

    Assign the monthly risk-free rate to t_bill_rate.
    Use the tangency.portfolio() function to calculate the efficient portfolio characteristics for both cases, that is, one where short sales is allowed, and one where short sales is not allowed. Assign the results to tangency_portfolio_short and tangency_portfolio_no_short.
    Study the plots that represent the weights of the stocks in both portfolios. The corresponding plot() functions are already provided.